USERDOMAIN check change for TFS Hosted Build
Sometime after mid November 2015, Microsoft changed their USERDOMAIN and
COMPUTERNAME Properties. As of 1/4/2016, my TFS hosted build was
building on COMPUTERNAME = BUILD2-0017, and USERDOMAIN is also
BUILD2-0017.
This update will accommodate these new property names. With the existing
code - with the new computer name - it thought the build was
TFSONPREMBUILD, and threw the Exception:
Microsoft.TeamFoundation.TeamFoundationServerUnauthorizedException:
TF30063: You are not authorized to access
https://{YourCollectionName}.visualstudio.com/DefaultCollection.